+++ foundation_stdlib/lib/vector.lua
@@ -9,10 +9,22 @@ end
 -- @namespace foundation.com.Vector2
 local vector2 = {}
 
+-- @type Vector2: {
+--   x: Number,
+--   y: Number,
+-- }
+
+-- @spec new(x: Number, y: Number): Vector2
 function vector2.new(x, y)
   return { x = x, y = y }
 end
 
+-- @spec copy(Vector2): Vector2
+function vector2.copy(v1)
+  return { x = v1.x, y = v1.y }
+end
+
+-- @spec unwrap(Any): (Number, Number)
 function vector2.unwrap(v)
   if type(v) == "table" then
     return v.x, v.y
@@ -22,55 +34,65 @@ function vector2.unwrap(v)
   error("expected a table or number")
 end
 
+-- @spec zero(): Vector2
 function vector2.zero()
   return vector2.new(0, 0)
 end
 
-function vector2.to_string(v1, seperator)
+-- @spec to_string(Vector2, seperator?: String): String
+function vector2.to_string(v2, seperator)
   seperator = seperator or ","
-  return v1.x .. seperator .. v1.y
+  return v2.x .. seperator .. v2.y
 end
 
-function vector2.floor(dest, v1)
-  dest.x = math.floor(v1.x)
-  dest.y = math.floor(v1.y)
+-- @spec floor(dest: Vector2, v2: Vector2): Vector2
+function vector2.floor(dest, v2)
+  dest.x = math.floor(v2.x)
+  dest.y = math.floor(v2.y)
   return dest
 end
 
-function vector2.round(dest, v1)
-  dest.x = math.floor(v1.x + 0.5)
-  dest.y = math.floor(v1.y + 0.5)
+-- @spec round(dest: Vector2, v2: Vector2): Vector2
+function vector2.round(dest, v2)
+  dest.x = math.floor(v2.x + 0.5)
+  dest.y = math.floor(v2.y + 0.5)
   return dest
 end
 
+-- @spec dot(Vector2, Vector2): Vector2
 function vector2.dot(v1, v2)
   return v1.x * v2.x + v1.y * v2.y
 end
 
+-- @spec add(dest: Vector2, Vector2, Vector2): Vector2
 function vector2.add(dest, v1, v2)
   dest.x = v1.x + v2.x
   dest.y = v1.y + v2.y
   return dest
 end
 
+-- @spec subtract(dest: Vector2, Vector2, Vector2): Vector2
 function vector2.subtract(dest, v1, v2)
   dest.x = v1.x + v2.x
   dest.y = v1.y + v2.y
   return dest
 end
 
+-- @spec multiply(dest: Vector2, Vector2, Vector2): Vector2
 function vector2.multiply(dest, v1, v2)
   dest.x = v1.x * v2.x
   dest.y = v1.y * v2.y
   return dest
 end
 
+-- @spec divide(dest: Vector2, Vector2, Vector2): Vector2
 function vector2.divide(dest, v1, v2)
   dest.x = v1.x / v2.x
   dest.y = v1.y / v2.y
   return dest
 end
 
+-- @spec idivide(dest: Vector2, Vector2, Vector2): Vector2
 function vector2.idivide(dest, v1, v2)
   dest.x = math.floor(v1.x / v2.x)
   dest.y = math.floor(v1.y / v2.y)
@@ -85,10 +107,23 @@ vector2.idiv = vector2.idivide
 -- @namespace foundation.com.Vector3
 local vector3 = {}
 
+-- @type Vector3: {
+--   x: Number,
+--   y: Number,
+--   z: Number,
+-- }
+
+-- @spec new(x: Number, y: Number, z: Number): Vector3
 function vector3.new(x, y, z)
   return { x = x, y = y, z = z }
 end
 
+-- @spec copy(Vector3): Vector3
+function vector3.copy(v1)
+  return { x = v1.x, y = v1.y, z = v1.z }
+end
+
+-- @spec new(Any): (Number, Number, Number)
 function vector3.unwrap(v)
   if type(v) == "table" then
     return v.x, v.y, v.z
@@ -98,15 +133,18 @@ function vector3.unwrap(v)
   error("expected a table or number")
 end
 
+-- @spec zero(): Vector3
 function vector3.zero()
   return vector3.new(0, 0, 0)
 end
 
+-- @spec to_string(Vector3, seperator?: String): String
 function vector3.to_string(v1, seperator)
   seperator = seperator or ","
   return v1.x .. seperator .. v1.y .. seperator .. v1.z
 end
 
+-- @spec floor(Vector3, Vector3): Vector3
 function vector3.floor(dest, v1)
   dest.x = math.floor(v1.x)
   dest.y = math.floor(v1.y)
@@ -114,6 +152,7 @@ function vector3.floor(dest, v1)
   return dest
 end
 
+-- @spec round(Vector3, Vector3): Vector3
 function vector3.round(dest, v1)
   dest.x = math.floor(v1.x + 0.5)
   dest.y = math.floor(v1.y + 0.5)
@@ -121,10 +160,12 @@ function vector3.round(dest, v1)
   return dest
 end
 
+-- @spec dot(Vector3, Vector3): Number
 function vector3.dot(v1, v2)
   return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z
 end
 
+-- @spec add(dest: Vector3, Vector3, Vector3): Vector3
 function vector3.add(dest, v1, v2)
   local v1x, v1y, v1z = vector3.unwrap(v1)
   local v2x, v2y, v2z = vector3.unwrap(v2)
@@ -134,6 +175,7 @@ function vector3.add(dest, v1, v2)
   return dest
 end
 
+-- @spec subtract(dest: Vector3, Vector3, Vector3): Vector3
 function vector3.subtract(dest, v1, v2)
   local v1x, v1y, v1z = vector3.unwrap(v1)
   local v2x, v2y, v2z = vector3.unwrap(v2)
@@ -143,6 +185,7 @@ function vector3.subtract(dest, v1, v2)
   return dest
 end
 
+-- @spec multiply(dest: Vector3, Vector3, Vector3): Vector3
 function vector3.multiply(dest, v1, v2)
   local v1x, v1y, v1z = vector3.unwrap(v1)
   local v2x, v2y, v2z = vector3.unwrap(v2)
@@ -152,6 +195,7 @@ function vector3.multiply(dest, v1, v2)
   return dest
 end
 
+-- @spec divide(dest: Vector3, Vector3, Vector3): Vector3
 function vector3.divide(dest, v1, v2)
   local v1x, v1y, v1z = vector3.unwrap(v1)
   local v2x, v2y, v2z = vector3.unwrap(v2)
@@ -161,6 +205,7 @@ function vector3.divide(dest, v1, v2)
   return dest
 end
 
+-- @spec idivide(dest: Vector3, Vector3, Vector3): Vector3
 function vector3.idivide(dest, v1, v2)
   local v1x, v1y, v1z = vector3.unwrap(v1)
   local v2x, v2y, v2z = vector3.unwrap(v2)
@@ -178,10 +223,24 @@ vector3.idiv = vector3.idivide
 -- @namespace foundation.com.Vector4
 local vector4 = {}
 
+-- @type Vector4: {
+--   x: Number,
+--   y: Number,
+--   z: Number,
+--   w: Number,
+-- }
+
+-- @spec new(x: Number, y: Number, z: Number, w: Number): Vector4
 function vector4.new(x, y, z, w)
   return { x = x, y = y, z = z, w = w }
 end
 
+-- @spec copy(Vector4): Vector4
+function vector4.copy(v1)
+  return { x = v1.x, y = v1.y, z = v1.z, w = v1.w }
+end
+
+-- @spec unwrap(Any): (Number, Number, Number, Number)
 function vector4.unwrap(v)
   if type(v) == "table" then
     return v.x, v.y, v.z, v.w
@@ -191,14 +250,18 @@ function vector4.unwrap(v)
   error("expected a table or number")
 end
 
+-- @spec zero(): Vector4
 function vector4.zero()
-  return vector4.new(0, 0, 0)
+  return vector4.new(0, 0, 0, 0)
 end
 
-function vector4.to_string(v1)
+-- @spec to_string(Vector4, seperator?: String): String
+function vector4.to_string(v1, seperator)
+  seperator = seperator or ","
   return v1.x .. "," .. v1.y .. "," .. v1.z .. "," .. v1.w
 end
 
+-- @spec floor(Vector4, Vector4): Vector4
 function vector4.floor(dest, v1)
   dest.x = math.floor(v1.x)
   dest.y = math.floor(v1.y)
@@ -207,6 +270,7 @@ function vector4.floor(dest, v1)
   return dest
 end
 
+-- @spec round(Vector4, Vector4): Vector4
 function vector4.round(dest, v1)
   dest.x = math.floor(v1.x + 0.5)
   dest.y = math.floor(v1.y + 0.5)
@@ -215,10 +279,12 @@ function vector4.round(dest, v1)
   return dest
 end
 
+-- @spec dot(Vector4, Vector4): Vector4
 function vector4.dot(v1, v2)
   return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z + v1.w * v2.w
 end
 
+-- @spec add(dest: Vector4, Vector4, Vector4): Vector4
 function vector4.add(dest, v1, v2)
   local v1x, v1y, v1z, v1w = vector4.unwrap(v1)
   local v2x, v2y, v2z, v2w = vector4.unwrap(v2)
@@ -229,6 +295,7 @@ function vector4.add(dest, v1, v2)
   return dest
 end
 
+-- @spec subtract(dest: Vector4, Vector4, Vector4): Vector4
 function vector4.subtract(dest, v1, v2)
   local v1x, v1y, v1z, v1w = vector4.unwrap(v1)
   local v2x, v2y, v2z, v2w = vector4.unwrap(v2)
@@ -239,6 +306,7 @@ function vector4.subtract(dest, v1, v2)
   return dest
 end
 
+-- @spec multiply(dest: Vector4, Vector4, Vector4): Vector4
 function vector4.multiply(dest, v1, v2)
   local v1x, v1y, v1z, v1w = vector4.unwrap(v1)
   local v2x, v2y, v2z, v2w = vector4.unwrap(v2)
@@ -249,6 +317,7 @@ function vector4.multiply(dest, v1, v2)
   return dest
 end
 
+-- @spec divide(dest: Vector4, Vector4, Vector4): Vector4
 function vector4.divide(dest, v1, v2)
   local v1x, v1y, v1z, v1w = vector4.unwrap(v1)
   local v2x, v2y, v2z, v2w = vector4.unwrap(v2)
@@ -259,9 +328,21 @@ function vector4.divide(dest, v1, v2)
   return dest
 end
 
+-- @spec idivide(dest: Vector4, Vector4, Vector4): Vector4
+function vector4.idivide(dest, v1, v2)
+  local v1x, v1y, v1z, v1w = vector4.unwrap(v1)
+  local v2x, v2y, v2z, v2w = vector4.unwrap(v2)
+  dest.x = math.floor(v1x / v2x)
+  dest.y = math.floor(v1y / v2y)
+  dest.z = math.floor(v1z / v2z)
+  dest.w = math.floor(v1w / v2w)
+  return dest
+end
+
 vector4.sub = vector4.subtract
 vector4.mul = vector4.multiply
 vector4.div = vector4.divide
+vector4.idiv = vector4.idivide
 
 foundation.com.Vector2 = vector2
 foundation.com.Vector3 = vector3